+++ cogs/anime.py
@@ -93,7 +93,7 @@ class Anime(commands.Cog):
     # RENAULT - SFW
     @commands.command(help = 'Renault. Renault.')
     async def renault(self, ctx):
-        await ctx.reply('https://cdn.discordapp.com/attachments/879488846453174282/895849572033781822/renault.png')
+        await ctx.reply(AnimeList.Renault())
 
 def setup(client):
     client.add_cog(Anime(client))